requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# TwelveLabs SDK
twelvelabs>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# Async HTTP client (for future API integrations)
httpx>=0.27.0

# Video processing (FFmpeg wrapper - will be added later)
# python-ffmpeg>=1.0.0
